Pause pump when supply pressure drops below 5 psi.
Based on observation the 5 psi mark is within two seconds of the 1 psi
mark. When using the 1 psi mark observed pressure spikes below zero to
-7 or -8 psi. Changing the cutoff to 5 psi will reduce the stress on the
system.

The reason for the spike so quickly is due to the pressurized buffer
tank. When the supply pressure drops below the pressure value of 5-7 psi
the only remainin water to drain is in the lines so pressure drops very
quickly.