Update pressure control limits.
Raise lower limit to 10psi after observing pump not shutting off when
buffer tank is depleted. Best hypothesis is air in the line after
changing filters caused low pressure not to reach sensor.

Widen normal operating pressure range. This reduces the frequency of
pump running and avoids short cycles in response to the pot filler being
opened.

This change corresponds with a filter change and reconnecting the second
14g storage tank.